Refactor to pytest and PluginMixin, merge into test_web.py
